box edit view (WIP / broken)
